Moved Percolator run-FDR orchestration into OspreySharp.FDR
* Introduced PercolatorEngine (OspreySharp.FDR) owning RunPercolatorFdr
  and the streaming dispatch; FirstJoinTask.RunPercolatorFdr is now a
  thin facade passing ctx.LogInfo + the PIN feature names.
* Moved PercolatorEntryBuilder into OspreySharp.FDR, parameterizing the
  feature count so it no longer reaches the Scoring/IO constants.
* Engine takes a log delegate + data, never PipelineContext. Byte-
  identical (Stellar golden mode 1 + 2).

See ai/todos/active/TODO-20260617_ospreysharp_debt_paydown_pr5.md
